+++ src/main/java/org/hydev/mcpm/client/arguments/parsers/InstallParser.java
@@ -4,8 +4,8 @@ import net.sourceforge.argparse4j.impl.Arguments;
 import net.sourceforge.argparse4j.inf.Namespace;
 import net.sourceforge.argparse4j.inf.Subparser;
 import org.hydev.mcpm.client.commands.controllers.InstallController;
-import org.hydev.mcpm.client.installer.presenter.InstallPresenter;
-import org.hydev.mcpm.client.installer.presenter.InstallResultPresenter;
+import org.hydev.mcpm.client.display.presenters.InstallPresenter;
+import org.hydev.mcpm.client.commands.presenters.InstallResultPresenter;
 import org.hydev.mcpm.client.search.SearchPackagesType;
 
 import java.util.function.Consumer;
