+++ src/main/java/org/hydev/mcpm/client/display/presenters/InstallPresenter.java
@@ -1,5 +1,6 @@
-package org.hydev.mcpm.client.installer.presenter;
+package org.hydev.mcpm.client.display.presenters;
 
+import org.hydev.mcpm.client.commands.presenters.InstallResultPresenter;
 import org.hydev.mcpm.client.installer.InstallResult;
 
 import java.util.function.Consumer;
